# perf: improve instantation speed by dynamic formatter generation