--- a/lib/countSpecifiers.js
+++ b/lib/countSpecifiers.js
@@ -1,0 +1,17 @@
+'use strict'
+
+const countSpecifiersRE = /%[sdifjoOc%]/g
+
+function countSpecifiers(message) {
+  let result = 0
+  message.replace(countSpecifiersRE, function (x) {
+    if (x !== '%%') {
+      result++
+    }
+    return x
+  })
+
+  return result
+}
+
+module.exports = countSpecifiers