--- a/lib/countSpecifiers.js
+++ b/lib/countSpecifiers.js
@@ -2,7 +2,7 @@
 
 const countSpecifiersRE = /%[sdifjoOc%]/g
 
-function countSpecifiers(message) {
+function countSpecifiers (message) {
   let result = 0
   message.replace(countSpecifiersRE, function (x) {
     if (x !== '%%') {