--- a/lib/formatFn.js
+++ b/lib/formatFn.js
@@ -1,16 +1,12 @@
 'use strict'
 
-const countSpecifiers = require("./countSpecifiers")
 const specifiersRE = /(%[sdifjoOc%])/g
 const inspect = require('util').inspect
+const countSpecifiers = require('./countSpecifiers')
 
 function formatFn(message) {
-  const amountOfPlaceholders = countSpecifiers(message)
 
-  if (amountOfPlaceholders === 0) {
-    message = message.replace(/%%/g, '%')
-    return () => message
-  }
+  const specifiersAmount = countSpecifiers(message)
 
   let fnBody = `
   function stringify(value) {
@@ -27,6 +23,16 @@ function formatFn(message) {
 
   const oOptions = { showHidden: true, showProxy: true }
 
+  function rest (args) {
+    let result = ''
+    let i = ${specifiersAmount}
+    const argsLength = args.length
+    while (i < args.length) {
+      result += ' ' + args[i++]
+    }
+    return result
+  }
+
   return function format (args) {
 
     const argsLen = args.length
@@ -41,7 +47,7 @@ function formatFn(message) {
       case '':
         break
       case '%%':
-        fnBody += ' + "%"'
+        fnBody += ' + (argsLen === 0 ? "%%" : "%")'
         break;
       case '%d':
         fnBody += ` + (
@@ -74,7 +80,11 @@ function formatFn(message) {
       case '%s':
         fnBody += ` + (
           ${argNum} >= argsLen && '%s' ||
-          args[${argNum}]
+          (
+            (typeof args[${argNum}] === 'bigint' && args[${argNum}].toString() + 'n') ||
+            (typeof args[${argNum}] === 'number' && args[${argNum}] === 0 && 1 / args[${argNum}] === -Infinity && '-0') ||
+            args[${argNum}]
+          )
         )`
         argNum++
         break;
@@ -106,9 +116,11 @@ function formatFn(message) {
     }
   }
 
+  fnBody += '+ rest(args)'
+
   fnBody += '}'
 
-  return new Function(fnBody)(inspect)
+  return new Function('inspect', fnBody)(inspect)
 }