--- a/lib/formatFn.js
+++ b/lib/formatFn.js
@@ -4,8 +4,7 @@ const specifiersRE = /(%[sdifjoOc%])/g
 const inspect = require('util').inspect
 const countSpecifiers = require('./countSpecifiers')
 
-function formatFn(message) {
-
+function formatFn (message) {
   const specifiersAmount = countSpecifiers(message)
 
   let fnBody = `
@@ -48,14 +47,14 @@ function formatFn(message) {
         break
       case '%%':
         fnBody += ' + (argsLen === 0 ? "%%" : "%")'
-        break;
+        break
       case '%d':
         fnBody += ` + (
           ${argNum} >= argsLen && '%d' ||
           args[${argNum}]
         )`
         argNum++
-        break;
+        break
       case '%i':
         fnBody += ` + (
           ${argNum} >= argsLen && '%i' ||
@@ -69,14 +68,14 @@ function formatFn(message) {
           parseInt(args[${argNum}], 10)
         )`
         argNum++
-        break;
+        break
       case '%f':
         fnBody += ` + (
           ${argNum} >= argsLen && '%f' ||
           parseFloat(args[${argNum}])
         )`
         argNum++
-        break;
+        break
       case '%s':
         fnBody += ` + (
           ${argNum} >= argsLen && '%s' ||
@@ -87,32 +86,32 @@ function formatFn(message) {
           )
         )`
         argNum++
-        break;
+        break
       case '%o':
         fnBody += ` + (
           ${argNum} >= argsLen && '%o' ||
           inspect(args[${argNum}], oOptions)
         )`
         argNum++
-        break;
+        break
       case '%O':
         fnBody += ` + (
           ${argNum} >= argsLen && '%O' ||
           inspect(args[${argNum}])
         )`
         argNum++
-        break;
+        break
       case '%j':
         fnBody += ` + (
           ${argNum} >= argsLen && '%j' ||
           stringify(args[${argNum}])
         )`
         argNum++
-        break;
+        break
       case '%c':
-        break;
+        break
       default:
-        fnBody += "+ " + JSON.stringify(messagePart)
+        fnBody += '+ ' + JSON.stringify(messagePart)
     }
   }
 
@@ -120,8 +119,7 @@ function formatFn(message) {
 
   fnBody += '}'
 
-  return new Function('inspect', fnBody)(inspect)
+  return new Function('inspect', fnBody)(inspect) // eslint-disable-line no-new-func
 }
-
 
 module.exports = formatFn