--- a/test/lib/countSpecifiers.test.js
+++ b/test/lib/countSpecifiers.test.js
@@ -24,6 +24,5 @@ test('countSpecifiers', t => {
       t.plan(1)
       t.equal(countSpecifiers(testCase), expected)
     })
-
   }
 })