--- a/test/lib/formtFn.test.js
+++ b/test/lib/formtFn.test.js
@@ -8,9 +8,18 @@ const circular = {}
 circular.circular = circular
 
 const testCases = [
-  ['%%', [], '%'],
+  ['%%', [], '%%'],
+  ['%% %s', [], '%% %s'],
+  ['%% %d', [2], '% 2'],
   ['no specifier', [], 'no specifier'],
   ['string %s', [0], 'string 0'],
+  ['string %s', [-0], 'string -0'],
+  ['string %s', [0n], 'string 0n'],
+  ['string %s', [Infinity], 'string Infinity'],
+  ['string %s', [-Infinity], 'string -Infinity'],
+  ['string %s', [-NaN], 'string NaN'],
+  ['string %s', [undefined], 'string undefined'],
+  ['%s', [{ toString: () => 'Foo' }], 'Foo'],
   ['integer %i', [0n], 'integer 0n'],
   ['integer %i', [Infinity], 'integer NaN'],
   ['integer %i', [-Infinity], 'integer NaN'],
@@ -24,25 +33,19 @@ const testCases = [
   ['float %f', [{}], 'float NaN'],
   ['json %j', [{}], 'json {}'],
   ['json %j', [circular], 'json [Circular]'],
-  ['%s:%s', ['foo'], 'foo:%s']
+  ['%s:%s', ['foo'], 'foo:%s'],
+  ['%s:%c', ['foo', 'bar'], 'foo:'],
+  ['%o', [{}], '{}'],
+  ['%O', [{}], '{}'],
+  ['1', [2, 3], '1 2 3']
 ]
 test('formatFn', t => {
   t.plan(testCases.length)
 
   for (const [testCase, args, expected] of testCases) {
     t.test(testCase, t => {
-      t.plan(1)
+      t.plan(2)
       t.equal(formatFn(testCase)(args), expected)
-    })
-  }
-})
-
-test('format compatibility', t => {
-  t.plan(testCases.length)
-
-  for (const [testCase, args] of testCases) {
-    t.test(testCase, t => {
-      t.plan(1)
       t.equal(formatFn(testCase)(args), format(testCase, ...args))
     })
   }