As I see that @trim21 is active in this repo, I wanted to suggest a change, which has maybe more impact than #100

This is just a poc. It is not complete regarding the serialization. 

To have some useful performance benchmark, i disabled stacktrace generation.

before:
```sh
uzlopak@Battlestation:~/workspace/fastify-org/fastify-error$ node benchmarks/instantiate.js
instantiate Error x 5,245,780 ops/sec ±0.37% (192 runs sampled)
instantiate FastifyError x 238,523,493 ops/sec ±0.07% (196 runs sampled)
instantiate FastifySpecialError x 20,584,535 ops/sec ±0.54% (194 runs sampled)
```
after:
```sh
uzlopak@Battlestation:~/workspace/fastify-org/fastify-error$ node benchmarks/instantiate.js
instantiate Error x 5,208,812 ops/sec ±0.22% (192 runs sampled)
instantiate FastifyError x 238,148,483 ops/sec ±0.08% (195 runs sampled)
instantiate FastifySpecialError x 144,541,760 ops/sec ±0.12% (194 runs sampled)
```

Obviously it gets slower if type casting is feature complete. But I assume, that it is still faster than calling util.format every time. 
Maybe it is useless, as stacktrace is always the biggest perf bottleneck. But in the case, somebody disables stacktrace on production services (which is i think anyway best practice) it shoiuld have an impact on perf.


<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [ ] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [ ] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)