@Uzlopak 2023-05-08
@ivan-tymoshenko 
This is maybe very similar to fast-json-stringify