@ivan-tymoshenko 2023-05-08
@Uzlopak sorry, I don't understand. Can you explain, please?