@Uzlopak 2023-05-09
new benchmarks:
before:
node benchmarks/instantiate.js
instantiate Error x 5,148,366 ops/sec ±0.28% (191 runs sampled)
instantiate FastifyError x 121,849,157 ops/sec ±0.50% (190 runs sampled)
instantiate FastifyError arg 1 x 14,033,361 ops/sec ±0.25% (194 runs sampled)
instantiate FastifyError arg 2 x 10,830,852 ops/sec ±0.33% (191 runs sampled)

after:
node benchmarks/instantiate.js
instantiate Error x 5,049,694 ops/sec ±0.26% (193 runs sampled)
instantiate FastifyError x 236,957,612 ops/sec ±0.09% (193 runs sampled)
instantiate FastifyError arg 1 x 83,758,765 ops/sec ±0.91% (185 runs sampled)
instantiate FastifyError arg 2 x 63,728,097 ops/sec ±2.10% (179 runs sampled)