@Uzlopak 2023-05-09
@ivan-tymoshenko 

I use now the message string as pattern to create the format function once instead of calling format(message, args) every time.