@Uzlopak 2023-05-09
With default stackTraceLimit

before:
node benchmarks/instantiate.js
instantiate Error x 322,998 ops/sec ±2.65% (137 runs sampled)
instantiate FastifyError x 231,124 ops/sec ±1.33% (188 runs sampled)
instantiate FastifyError arg 1 x 171,481 ops/sec ±1.44% (182 runs sampled)

after:
node benchmarks/instantiate.js
instantiate Error x 304,488 ops/sec ±3.00% (138 runs sampled)
instantiate FastifyError x 224,038 ops/sec ±1.68% (187 runs sampled)
instantiate FastifyError arg 1 x 188,652 ops/sec ±1.69% (188 runs sampled)
instantiate FastifyError arg 2 x 180,878 ops/sec ±1.80% (176 runs sampled)

In conclusio: 

It is signigifcant faster if we set stackTraceLimit to 0. If we use default stackTraceLimit, we are not definetly not faster.
instantiate FastifyError arg 2 x 188,166 ops/sec ±1.61% (181 runs sampled)