@jsumners 2023-05-10
I think we already had a discussion and determined that we do not want to omit stack traces. They are the purpose of the error.