@trim21 2023-05-10
> What is the goal of the change? What are are you trying to achieve?

improve string formatting perf. disabling stack trace is would make it easier to focus on the perf of string formatting, not actually disable it for perf.