@Uzlopak 2023-05-11
I think, that in production Environment disabling stacktraces is a viable decision. In development using stacktraces make obviously sense.