@jsumners 2023-05-11
Sure, it's a viable decision. But I would recommend against it. Without the stack trace you are left with very little information to determine the cause of issues.